requirements.txt
Flask==3.0.0
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.5.1
google-generativeai>=0.7.2
gunicorn==21.2.0
